Prevent unlimited growth of AutoConfiguredCompositeMeterRegistry
Update `OnlyOnceLoggingDenyMeterFilter` to directly implement the
maximum number of allowable tag values. With this change in place
we can remove calls to `MeterFilter.maximumAllowableTags`. This also
allows `MeterRegistryPostProcessor` to have a stronger signal about
the filters that it should also apply to the
`AutoConfiguredCompositeMeterRegistry` since it can now do a standard
`instanceof` check.

Fixes gh-47285