Tolerate race condition in httpRequestsAreTimed
Tomcat 11 will [1] send a response as soon as Content-Length bytes has
been written. This initiates a race between the timer being registered
on the server side and the test receiving the response and looking for
the timer.

Rather than sleeping for a fix period of time, we now use Awaitility
to await the availability of the timer.

Closes gh-48049

[1] https://github.com/apache/tomcat/commit/69eff83577f7c00cbaaca9384ab4b1989f516797