Fix bug in SslMeterBinder when binding without registered bundles
Before this commit the meter registries were derived from the created
gauges. If the SslMeterBinder has been bound to a MeterRegistry without
any bundles, then no gauges are created. If a SslBundle is then
dynamically added, onBundleChange is called with the new bundle, but the
list of meter registries is empty (because we have no gauges). The
effect is that the newly registered bundle has no metrics.

Closes gh-48180