Improve Jackson configuration for the actuator endpoints to include package private constructors.
Jackson 3 does not detected package private constructors anyore in case they have more than one argument. This prevents the Restclient / Resttemplate and their test derivates deserializing metrics endpoints as they could do in Spring Boot 3.x with Jackson 2.

This changes the isolated mapper to the behaviour of Jackson 2, so that the mapping works out of the box again.

Signed-off-by: Michael Simons <michael@simons.ac>