Relax conditions to create RestClient infrastructure
Previously, the necessary infrastructure to create a RestClient was only
configured in a servlet-based application or in a reactive application
if virtual threads are enabled.

While this extra care was important in
Spring Boot 4 as the aut-configuration is always available, this is no
longer the case with Spring Boot 4. Indeed, an explicit module has to be
added to the classpath now.

This commit therefore relaxes the condition. If the module has been
added, then the infrastructure is auto-configured.

Closes gh-48308