Revisit condition to skip AOT processing of tests
This commit stop skipping AOT processing of tests when tests execution
is skipped. With this commit, only the presence of maven.test.skip makes
the processing to skip.

This harmonize the behavior with other standard plugins that do react
to the property. For instance, the compiler plugin will not compile the
tests if that property is set, but would do so if skipTests was
provided.

Closes gh-46929